consolidate protocol v0.0.1
